Convert HTML to proper Jekyll's _data
Ideally, we should use Jekyll's YML Data feature when dealing
data entries, so it will be more editable and readable.
Current site still using HTML for storing company information.

This commit will change this format

    <div>
      <a href="http://www.alodokter.com">
        <image src="images/logo_alodokter.png">
      </a>
      <h3>Alodokter.com</h3>
      <p>Alodokter.com adalah situs kesehatan No. 1 di Indonesia yang menggunakan Ruby sebagai back-end untuk App dan Web.</p>
    </div>

Into this

    - name: Alodokter.com
      url: http://www.alodokter.com
      image_url: images/logo_alodokter.png
      description: Alodokter.com adalah situs kesehatan No. 1 di Indonesia yang menggunakan Ruby sebagai back-end untuk App dan Web.

This commit also updated the README instruction on
how to add more company to the site.